requirements.txt
requests==2.5.1
python-dateutil==2.4.0
six==1.9.0
